# Remove timer on UndoRedoHelper and manually manage undo/redo snapshots a bit more