## What does the pull request do?

Improve the undo/redo system for `TextBox` by removing the need for `UndoRedoHelper` to use a 1-second timer to store things. (The timer was causing weird issues that were hard to reproduce on my end -- sometimes I couldn't undo or redo as I expected.) The `UndoRedoHelper` class was only used by `TextBox`, nor was it `public`, so this shouldn't have unintended side effects outside of `TextBox`. Importantly, this will make undo redo **consistent** because it no longer relies on a timer.

This also (in my opinion) improves undo/redo for `TextBox` by manually managing the stack a little bit more to be more intuitive: snapshot on space, snapshot in the middle of long words (every 7 changes; 7 was arbitrary), things like that.

I had trouble getting the initial state captured (e.g. from `` to `a`), so I added a flag to manually get that edge case, which I don't like; perhaps someone can see a better way to get that snapshot in. I'm also not sure this will work well for non-English languages, but I think (??) it should work OK since we snapshot in the middle of long words. **It would be helpful to me if someone could test this.**

I'm guessing this should have some sort of perf improvement, since we don't have a timer going off every 1 second for every textbox that's been created, but I don't know how to show this nicely in this PR with pretty graphs or similar.

By snapshotting before an undo is about to occur, we removed the need to snapshot after a text change (e.g. a snapshot -> handle text change -> another snapshot).

This also fixes a bug where the `UndoRedoHelper.Snapshot()` func to `_currentNode.Value.Equals(current)` did not properly call the `public bool Equals(UndoRedoState other)`. Not calling the right `Equals` func could cause an undo operation to snapshot the same text (but with a different `CaretIndex` as a different item to snapshot rather than a different one.

In summary, this PR:

* Removes need for undo redo helper snapshot timer that happened every 1s for every `TextBox` (a perf improvement at some level)
* Removes extra `Snapshot()` calls for undo/redo that are no longer necessary due to a snapshot that takes place before and undo operation
* Fixes `UndoRedoState.Equals` use
* Makes undo/redo a little nicer to use and much more consistent (because not relying on a timer)

### Before

(Note the weird undo to the middle of `hello`, probably caused by the timer)

![before](https://user-images.githubusercontent.com/5092399/120937170-ef4dfa80-c6d9-11eb-8a8a-5964baf82c98.gif)

### After

![after](https://user-images.githubusercontent.com/5092399/120937127-ac8c2280-c6d9-11eb-93b6-733a5d25afea.gif)

## What is the current behavior?

The `UndoRedoHelper` class uses a timer every 1 second to snapshot the text.

## What is the updated/expected behavior with this PR?

Got rid of the timer and manually managed things.

## How was the solution implemented (if it's not obvious)?

It's fairly straight forward. Nothing magical, really.

## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Documentation with user documentation

## Breaking changes

Undo/redo functions differently now, but I think it's an improvement.

## Obsoletions / Deprecations
<!--- Obsolete and Deprecated attributes on APIs MUST only be included when discussed with Core team. @grokys, @kekekeks & @danwalmsley -->

## Fixed issues
Fixes #3370