@Deadpikle 2021-06-25
Tried to run some benchmarks.

```csharp
[Benchmark]
public TextBox InitializeTextBox()
{
    using (UnitTestApplication.Start(TestServices.MockThreadingInterface))
    {
        return new TextBox();
    }
}
```

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.410
  [Host]     : .NET Core 3.1.16 (CoreCLR 4.700.21.26205, CoreFX 4.700.21.26205), X64 RyuJIT
  DefaultJob : .NET Core 3.1.16 (CoreCLR 4.700.21.26205, CoreFX 4.700.21.26205), X64 RyuJIT
```

#### Before (master at beaf783b7)

|            Method |         Mean |        Error |       StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |-------------:|-------------:|-------------:|-------:|------:|------:|----------:|
| InitializeTextBox | 159,134.0 ns | 17,300.34 ns | 49,358.84 ns |      - |     - |     - |   16328 B |

#### With this change (the branch is a little old but oh well)

|            Method |         Mean |        Error |       StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |-------------:|-------------:|-------------:|-------:|------:|------:|----------:|
| InitializeTextBox | 125,612.0 ns | 10,671.53 ns | 31,465.26 ns |      - |     - |     - |   13656 B |

So, a small improvement, I suppose.