@maxkatz6 2021-08-06
Removing "backport candidate" just to be safe about it. Keeping in mind that this PR does not adds any functionality, but optimizes existing.